Began Implementing the Linked List!
Created a linked list to organize and distribute information regarding the coordinates of each body segment in the display. Eventually these coordinates will be used to project the entire snake as it grows, and to determine if the snake crosses its own path and bumps into one of its segment which has to be detected because then the snake would have to die and result in a "game over".